more imrovements in business logic

+++ internal/service/transport.go
@@ -10,7 +10,6 @@ func (s *Service) CreateTransport(ownerId int, req models.TransportRequest) (int
 	if err != nil {
 		return -1, err
 	}
-
 	if !exists {
 		return -1, custom_errors.ErrIdNotFound
 	}
@@ -18,11 +17,9 @@ func (s *Service) CreateTransport(ownerId int, req models.TransportRequest) (int
 	if err := validateTransportType(req.TransportType); err != nil {
 		return -1, err
 	}
-
 	if err := validateTransportProperties(req.Model, req.Color, req.Description, req.Identifier); err != nil {
 		return -1, err
 	}
-
 	if req.Latitude < 0 || req.Longitude < 0 || req.MinutePrice < 0 || req.DayPrice < 0 {
 		return -1, custom_errors.ErrInvalidParams
 	}
@@ -35,7 +32,6 @@ func (s *Service) GetTransport(transportId int) (models.TransportResponse, error
 	if err != nil {
 		return models.TransportResponse{}, err
 	}
-
 	if !exists {
 		return models.TransportResponse{}, custom_errors.ErrIdNotFound
 	}
@@ -48,7 +44,6 @@ func (s *Service) UpdateTransport(userId int, transportId int, req models.Transp
 	if err != nil {
 		return err
 	}
-
 	if !exists {
 		return custom_errors.ErrIdNotFound
 	}
@@ -57,7 +52,6 @@ func (s *Service) UpdateTransport(userId int, transportId int, req models.Transp
 	if err != nil {
 		return err
 	}
-
 	if !exists {
 		return custom_errors.ErrIdNotFound
 	}
@@ -66,7 +60,6 @@ func (s *Service) UpdateTransport(userId int, transportId int, req models.Transp
 	if err != nil {
 		return err
 	}
-
 	if userId != ownerId {
 		return custom_errors.ErrAccessDenied
 	}
@@ -79,7 +72,6 @@ func (s *Service) DeleteTransport(userId int, transportId int) error {
 	if err != nil {
 		return err
 	}
-
 	if !exists {
 		return custom_errors.ErrIdNotFound
 	}
@@ -88,7 +80,6 @@ func (s *Service) DeleteTransport(userId int, transportId int) error {
 	if err != nil {
 		return err
 	}
-
 	if !exists {
 		return custom_errors.ErrIdNotFound
 	}
@@ -97,9 +88,16 @@ func (s *Service) DeleteTransport(userId int, transportId int) error {
 	if err != nil {
 		return err
 	}
-
 	if userId != ownerId {
 		return custom_errors.ErrAccessDenied
+	}
+
+	has, err := s.repo.CheckTransportIdHasActiveRents(transportId)
+	if err != nil {
+		return err
+	}
+	if has {
+		return custom_errors.ErrCanNotDelete
 	}
 
 	return s.repo.DeleteTransport(transportId)
